Move fake CAN reader to a separate utils lib
