+++ build.sh
@@ -7,7 +7,7 @@ set -e
 
 mkdir -p build 
 cd build
-cmake .. -DBUILD_EXAMPLES=ON -DBUILD_TESTS=ON -DBUILD_INTEGRATION_TESTS=ON
+cmake .. -DBUILD_EXAMPLES=ON -DBUILD_TESTS=ON -DBUILD_UTILS=ON -DBUILD_INTEGRATION_TESTS=ON
 make -j$(nproc)
 
 echo ""
