+++ rebuild.sh
@@ -3,7 +3,7 @@
 rm -rf build
 mkdir -p build 
 cd build
-cmake .. -DBUILD_EXAMPLES=ON -DBUILD_TESTS=ON -DBUILD_INTEGRATION_TESTS=ON
+cmake .. -DBUILD_EXAMPLES=ON -DBUILD_TESTS=ON -DBUILD_UTILS=ON -DBUILD_INTEGRATION_TESTS=ON
 make -j8
